wishlist divided into empty and not
