+++ Selenium_OpenCart/Pages/Body/WishListPage/EmptyWishListPage.cs
@@ -3,11 +3,15 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using OpenQA.Selenium;
 
 namespace Selenium_OpenCart.Pages.Body.WishListPage
 {
-    class EmptyWishListPage
+    class EmptyWishListPage : WishListPage
     {
+        public EmptyWishListPage(IWebDriver driver) : base(driver)
+        {
 
+        }
     }
 }
